QcomModulePkg : Update bootdevice string in the device tree
Currently, we don't update the device tree if the dynamically read
bootdevice string does not matches the length of the bootdevice
string mentioned in the device tree.

This patch fixes issue if the dynamically read bootdevice string
lenght is less than the one there in the device tree and it is
more then we will just continue without updating.

Change-Id: I6c311a2c7f1fe3f765f3ecff1146bef22ecd6ba4