QcomModulePkg: Fix reboot bootloader issue
In case if swap image store on partition from
previous boot and user issue the command reboot
bootloader, ABL first checks for swap image and load
swap image into the ram and gives control to it, in
doing so it skip checking the boot reason that failed
device to go into the fastboot mode.
Fix this by first checking boot reason.

Change-Id: Ic2038d680c2a51dc183079485ac630548ceacb17
Signed-off-by: Atul Raut <araut@codeaurora.org>