QcomModulePkg: Increase memory allocations on LE NAND and EMMC targets
When flashing images in fastboot, we only allocate 85% of the largest
contiguous piece of memory. However, this is not sufficient for LE NAND
and EMMC devices with lower DDR sizes, so this change increases it to 90%.
This avoids sparsing of ubifs images, which is not supported by the
fastboot host. This requires that we decrease the usb buffer size though
so we don't use too much heap space.

Change-Id: If8a5405d2124d6325ae460ed49a9040049455507