libavb: Split boot image to load and verify in parallel.
To reduce the boot time, it is required to divide boot image
load and verification in to two parts which can work
independently.

Boot image is divided in to two parts. Load and verification
of first part can be done by one cpu.
while verification is being done by first cpu.second cpu will
load the second part of the memory.

Hence verification of first part will go in parallel with
load of second part.

Change-Id: Id55125eed9b3758d9268be78fdd1f374cb287440
Signed-off-by: Santosh Dronamraju <sdronamr@codeaurora.org>